+++ frontend/src/components/flights/EditFlightDetailsDialog.tsx
@@ -242,6 +242,8 @@ export default function EditFlightDetailsDialog({
                                 }}>
                                     <TimePicker
                                         ampm={false}
+                                        ampmInClock={false}
+                                        timeSteps={{minutes: 1}}
                                         label={t("TAKE_OFF_TIME")}
                                         value={takeOffat ? moment(takeOffat) : null}
                                         onChange={(newValue: moment.Moment | null) => {
@@ -275,6 +277,8 @@ export default function EditFlightDetailsDialog({
                                 }}>
                                     <TimePicker
                                         ampm={false}
+                                        ampmInClock={false}
+                                        timeSteps={{minutes: 1}}
                                         label={t("TOW_RELEASE_TIME")}
                                         value={towReleaseAt ? moment(towReleaseAt) : null}
                                         onChange={(newValue: moment.Moment | null) => {
@@ -308,6 +312,10 @@ export default function EditFlightDetailsDialog({
                                 }}>
                                     <TimePicker
                                         ampm={false}
+                                        ampmInClock={false}
+                                        timeSteps={{
+                                            minutes: 1
+                                        }}
                                         label={t("LANDING_TIME")}
                                         value={landingAt ? moment(landingAt) : null}
                                         onChange={(newValue: moment.Moment | null) => {