+++ RELEASE-NOTES.md
@@ -15,6 +15,7 @@
 - Changes the naming of the `Solo Flight?` button in the flight creation dialog for members flight. [Issue #190](https://github.com/ohadch/gliding-action-page-v3/issues/190)
 - Updates the flight states colors: blue for draft, yellow for tow, red for inflight, and green for landed.
 - Turns the time pickers from am/pm to 24h format.
+- Configures minutes steps to 1 min instead of 5.
 
 ### Enhancements